// src/lib/types/PieceSet.ts
export interface PieceSet {
    name: string;
    pieces: {
        wK: string;
        wQ: string;
        wR: string;
        wB: string;
        wN: string;
        wP: string;
        bK: string;
        bQ: string;
        bR: string;
        bB: string;
        bN: string;
        bP: string;
    };
    board: string;
    square: string;
}